backend/requirements.txt
asn1crypto==1.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9; python_version >= '3'
click==8.0.3
cryptography==36.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
flask==2.0.2
greenlet==1.1.2; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
gunicorn==20.1.0
idna==3.3; python_version >= '3'
itsdangerous==2.0.1
jinja2==3.0.3
markupsafe==2.0.1
marshmallow==3.14.1
minio==7.1.2
nanoid==2.0.0
pg8000==1.23.0
pycparser==2.21
pygments==2.10.0
pyjwt==2.3.0
python-dotenv==0.19.2
python-http-client==3.3.4
requests==2.26.0
scramp==1.4.1
sendgrid==6.9.3
six==1.16.0
sqlalchemy==1.4.28
starkbank-ecdsa==2.0.3
ua-parser==0.10.0
urllib3==1.26.7
werkzeug==2.0.2
